Migrate the DesignComposeDefinition to dc_bundle
